+++ traces/lib/config/router.dart
@@ -13,24 +13,18 @@ import '../screens/notes/screens/note_detail_view.dart';
 import '../screens/notes/screens/note_page.dart';
 import '../screens/profile/bloc/profile/bloc.dart';
 import '../screens/profile/profile_page.dart';
-import '../screens/profile/repository/__firebase_profile_repository.dart';
 import '../screens/settings/bloc/settings_bloc.dart';
-import '../screens/settings/repository/firebase_appSettings_repository.dart';
 import '../screens/settings/settings_page.dart';
 import '../screens/settings/themes_settings_view.dart';
 import '../screens/trips/bloc/trips_bloc.dart';
-import '../screens/trips/repository/firebase_trips_repository.dart';
 import '../screens/trips/start_planning/bloc/startplanning_bloc.dart';
 import '../screens/trips/start_planning/start_planning_view.dart';
 import '../screens/trips/tripdetails/bloc/tripdetails_bloc.dart';
 import '../screens/trips/tripdetails/tripdetails_view.dart';
 import '../screens/trips/trips_page.dart';
-import '../screens/visas/bloc/entry_exit/visa_entry_bloc.dart';
 import '../screens/visas/bloc/visa/visa_bloc.dart';
 import '../screens/visas/bloc/visa_details/visa_details_bloc.dart';
 import '../screens/visas/bloc/visa_tab/visa_tab_bloc.dart';
-import '../screens/visas/entry_exit_details_view.dart';
-import '../screens/visas/repository/firebase_visas_repository.dart';
 import '../screens/visas/visa_details_view.dart';
 import '../screens/visas/visa_edit_view.dart';
 import '../screens/visas/visas_page.dart';
@@ -144,21 +138,7 @@ class RouteGenerator {
             );
           }
           return _errorRoute();
-        }
-      case visaAddEntryRoute:
-        {
-          if (args is String) {
-            return MaterialPageRoute(
-              builder: (_) => BlocProvider<VisaEntryBloc>(
-                create: (context) => VisaEntryBloc(
-                    /*visasRepository:
-                        FirebaseVisasRepository()*/) /*..add(args != '' ? EditVisaClicked(args) : NewVisaMode())*/,
-                child: VisaEntryDetailsView(/*entryId: args*/),
-              ),
-            );
-          }
-          return _errorRoute();
-        }
+        }      
       case tripsRoute:
         return MaterialPageRoute(
           builder: (_) => BlocProvider<TripsBloc>(
