+++ traces/lib/screens/visas/bloc/visa_details/visa_details_state.dart
@@ -2,8 +2,7 @@ part of 'visa_details_bloc.dart';
 
 class VisaDetailsState {
   Visa? visa;
-  final Group? familyGroup;
-  final List<EntryExit>? entryExits;
+  final Group? familyGroup; 
   final StateStatus status;
   final StateMode mode;  
   final String? errorMessage;
@@ -11,8 +10,7 @@ class VisaDetailsState {
 
   VisaDetailsState(
       {required this.visa,     
-      required this.familyGroup,
-      required this.entryExits,
+      required this.familyGroup,      
       required this.status,
       required this.mode,      
       this.errorMessage, this.activeTab});
@@ -20,8 +18,7 @@ class VisaDetailsState {
   factory VisaDetailsState.empty() {
     return VisaDetailsState(
         visa: null,       
-        familyGroup: null,
-        entryExits: null,
+        familyGroup: null,        
         status: StateStatus.Empty,
         mode: StateMode.View,
         errorMessage: "",
@@ -31,8 +28,7 @@ class VisaDetailsState {
   factory VisaDetailsState.loading() {
     return VisaDetailsState(
         visa: null,       
-        familyGroup: null,
-        entryExits: null,
+        familyGroup: null,       
         status: StateStatus.Loading,
         mode: StateMode.View,
         errorMessage: "",
@@ -48,20 +44,17 @@ class VisaDetailsState {
         familyGroup: members,
         status: StateStatus.Empty,
         mode: StateMode.Edit,        
-        errorMessage: "", entryExits: null,
+        errorMessage: "",
         activeTab: 0);
   }
 
   factory VisaDetailsState.success(
       {Visa? visa,
-      VisaSettings? settings,
-      UserSettings? userSettings,
-      Group? members,
-      List<EntryExit>? entryExits}) {
+      VisaSettings? settings,      
+      Group? members}) {
     return VisaDetailsState(
         visa: visa,        
-        familyGroup: members,
-        entryExits: entryExits,
+        familyGroup: members,       
         status: StateStatus.Success,
         mode: StateMode.View,
         errorMessage: "", activeTab: 0);
@@ -69,14 +62,12 @@ class VisaDetailsState {
 
   factory VisaDetailsState.failure(
       {Visa? visa,      
-      Group? members,
-      List<EntryExit>? entryExits,
+      Group? members,    
       bool? autovalidate,
       String? error}) {
     return VisaDetailsState(
         visa: visa,        
-        familyGroup: members,
-        entryExits: entryExits,
+        familyGroup: members,    
         status: StateStatus.Error,
         mode: StateMode.View,        
         errorMessage: error, activeTab: 0);
@@ -84,15 +75,13 @@ class VisaDetailsState {
 
   VisaDetailsState copyWith(
       {final Visa? visa,      
-      final Group? members,
-      final List<EntryExit>? entryExits,
+      final Group? members,    
       final StateStatus? status,
       final StateMode? mode,      
       String? errorMessage, int? activeTab}) {
     return VisaDetailsState(
         visa: visa ?? this.visa,        
-        familyGroup: members ?? this.familyGroup,
-        entryExits: entryExits ?? this.entryExits,
+        familyGroup: members ?? this.familyGroup,      
         status: status ?? this.status,
         mode: mode ?? this.mode,       
         errorMessage: errorMessage ?? this.errorMessage,
@@ -101,15 +90,14 @@ class VisaDetailsState {
   }
 
   VisaDetailsState update(
-      {Visa? visa,      
-      List<EntryExit>? entryExits,
+      {Visa? visa,     
+    
       StateStatus? status,
       StateMode? mode,      
       String? errorMessage, int? activeTab}) {
     return copyWith(
         visa: visa,        
-        members: familyGroup,
-        entryExits: entryExits,
+        members: familyGroup,      
         status: status,
         mode: mode,        
         errorMessage: errorMessage,
