+++ traces/lib/screens/visas/widgets/visa_delete_alert.dart
@@ -1,12 +1,12 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:traces/constants/color_constants.dart';
-import 'package:traces/constants/route_constants.dart';
-import 'package:traces/screens/visas/bloc/visa_details/visa_details_bloc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
-import 'package:traces/screens/visas/model/visa.model.dart';
-import 'package:traces/widgets/widgets.dart';
+
+import '../../../constants/color_constants.dart';
+import '../../../constants/route_constants.dart';
+import '../../../widgets/widgets.dart';
+import '../bloc/visa_details/visa_details_bloc.dart';
+import '../model/visa.model.dart';
 
 class VisaDeleteAlert extends StatelessWidget {
   final Visa? visa;
